Add styles for header section projects details
What solved
[X] Add the image style for the image and data small/desk resolutions
[X] Add new chip component small/desk
